Acceptance testing (#3)
* add first acceptance test for user sign in

* add user model

* add devise gem for auth

* add devise for user, config into app

* first acceptanse test passed

* both acceptance test for sign in passed

* add Authenticated user asks a question test

* add Unauthenticated user tries to ask a question test

* add Authenticated user asks a question with errors test

* add helper for failed controllers tests

* complete creating question story

* add Unauthenticated user can see list of questions test

* user can see list of questions

* User can see question page with answers

* refactor tests

* User on the question page can write answer

* User can log out

* User can sign up

* add association for question and user

* User can delete only his question

* in writing delete answer test

* controller tests pass

* some refactoring

* fix instance variable to call helper

* add tests for author_of method

* refactored factories

* add let for controllers tests

* create list of questions for index test

* add tests for delete answer

* fix tests for delete answer

* fix tests for delete questions

* fix test for create question

* update controllers test and macros

* rubocop fixes

* refactoring after CR

+++ app/controllers/questions_controller.rb
@@ -1,4 +1,5 @@
 class QuestionsController < ApplicationController
+  before_action :authenticate_user!, except: %i[index show]
   helper_method :question
 
   def index
@@ -12,9 +13,9 @@ class QuestionsController < ApplicationController
   def edit; end
 
   def create
-    @question = Question.new(question_params)
-    if @question.save
-      redirect_to @question
+    question = current_user.questions.build(question_params)
+    if question.save
+      redirect_to questions_path, notice: 'Your question successfully created.'
     else
       render :new
     end
@@ -29,8 +30,8 @@ class QuestionsController < ApplicationController
   end
 
   def destroy
-    question.destroy
-    redirect_to questions_path
+    question.destroy if current_user.author_of?(question)
+    redirect_to questions_path, notice: 'The question is successfully deleted.'
   end
 
   private
